% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatePeak.R
\name{AnnotatePeak2}
\alias{AnnotatePeak2}
\title{AnnotatePeak2}
\usage{
AnnotatePeak2(input.file.dir, input.file.pattern, index.file, output.file.dir,
  genome)
}
\arguments{
\item{input.file.dir}{}

\item{input.file.pattern}{}

\item{index.file}{}

\item{output.file.dir}{}

\item{genome}{}
}
\description{
AnnotatePeak2
}
\examples{

input.file.dir="/media/H_driver/2016/Danny/Danny_chip_PeakCall/"
input.file.pattern="*macs142_peaks.bed"
output.file.dir="/media/H_driver/2016/Danny/Danny_chip_PeakCall/"

AnnotatePeak(input.file.dir,input.file.pattern,8,output.file.dir,genome="Mm")
AnnotatePeak(input.file.dir,input.file.pattern,7,output.file.dir,genome="Hs")



}
