% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipSeq.R
\name{AnnotatePeak3}
\alias{AnnotatePeak3}
\title{AnnotatePeak3}
\usage{
AnnotatePeak3(input.file.dir, output.file.dir, genome)
}
\arguments{
\item{input.file.dir}{}

\item{output.file.dir}{}

\item{genome}{}
}
\description{
AnnotatePeak3
}
\examples{

input.file.dir="/media/H_driver/2016/Danny/Danny_chip_PeakCall/"
output.file.dir="/media/H_driver/2016/Danny/Danny_chip_PeakCall/"

AnnotatePeak3(input.file.dir,output.file.dir,genome="Mm")
AnnotatePeak3(input.file.dir,output.file.dir,genome="Hs")



}
