% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipSeq.R
\name{AnntationUsingChipSeeker}
\alias{AnntationUsingChipSeeker}
\title{AnntationUsingChipSeeker}
\usage{
AnntationUsingChipSeeker(dir.name, input.file.pattern, out.dir.name,
  txdb = c("hg19", "hg38"), DD, distanceToTSS_cutoff = 5000,
  assignGenomicAnnotation = TRUE, AP = c("Promoter", "5UTR", "3UTR", "Exon",
  "Intron", "Downstream", "Intergenic"))
}
\arguments{
\item{input.file.pattern}{: input file pattern}

\item{out.dir.name}{output file directory}

\item{txdb}{Annotation databased to be used("hg19","hg38")}

\item{DD}{distance definition around TSS}

\item{dir.name:}{path for bed files}
}
\description{
Use ChipSeeker to annotate data
}
\examples{

dir.name="/Volumes/Bioinformatics$/2017/DannyNewData/BindDiff/common_peaks_bed"
input.file.pattern="common_peaks.bed"
out.dir.name="/Volumes/Bioinformatics$/2017/DannyNewData/AnnotationNew"
txdb="hg19"
DD=5000

AnntationUsingChipSeeker(dir.name,input.file.pattern,out.dir.name,txdb=txdb,DD,distanceToTSS_cutoff=5000, AP=c("Promoter","Intron"))

AnntationUsingChipSeeker(dir.name,input.file.pattern,out.dir.name,txdb=txdb,DD,distanceToTSS_cutoff=5000,AP=c("Promoter"))

AnntationUsingChipSeeker(dir.name,input.file.pattern,out.dir.name,txdb=txdb,DD,distanceToTSS_cutoff=5000,AP=c("Intron"))
  
}
