% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFileSortIndexVisualization.R
\name{BamFileSortIndexVisualization2}
\alias{BamFileSortIndexVisualization2}
\title{R -e 'library(ChipSeq);library(ThreeUTR);ChipSeq:::BamFileSortIndexVisualization2("/scratch/projects/bbc/Project/Danny_chip2/Alignment/BWA","marked.bam",/scratch/projects/bbc/aiminy_project/DannyNewNgsPlot","Hs")'}
\usage{
BamFileSortIndexVisualization2(input, genome)
}
\description{
R -e 'library(ChipSeq);library(ThreeUTR);ChipSeq:::BamFileSortIndexVisualization2("/scratch/projects/bbc/Project/Danny_chip2/Alignment/BWA","marked.bam",/scratch/projects/bbc/aiminy_project/DannyNewNgsPlot","Hs")'
}
