% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-bindist.R
\docType{class}
\name{bindist-class}
\alias{bindist-class}
\alias{bindist}
\alias{bindist-method}
\alias{$,bindist-method}
\alias{$<-,bindist-method}
\title{Class \code{"bindist"}}
\description{
An S4 class representing the distribution of binding sites relative to
genomic features (e.g., TSS or gene end). Objects of this class store
fixed-width bin counts and positions for use in permutation testing to
generate background pools that match the observed binding distribution.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("bindist", counts="integer", mids="integer",
halfBinSize="integer", bindingType="character", featureType="character")}.
}

\seealso{
\link{preparePool}, \link{peakPermTest}
}
\keyword{classes}
