% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utilities-globals.R
\name{ChIPpeakAnno-globals}
\alias{ChIPpeakAnno-globals}
\alias{setChIPpeakAnnoGenome}
\alias{getChIPpeakAnnoGenome}
\alias{setChIPpeakAnnoTxDb}
\alias{getChIPpeakAnnoTxDb}
\alias{setChIPpeakAnnoEnsDb}
\alias{getChIPpeakAnnoEnsDb}
\alias{setChIPpeakAnnoSpecies}
\alias{getChIPpeakAnnoSpecies}
\alias{getTargetGenome}
\alias{setChIPpeakAnnoGlobals}
\title{Global Options for ChIPpeakAnno}
\description{
Internal functions to set and retrieve global options for ChIPpeakAnno,
including default genome assembly, species, TxDb, and EnsDb objects.
These options provide a convenient way to set defaults that are used across
multiple functions, while still allowing per-function overrides.
}
\details{
Most functions in this file are internal and not exported. Users should use
\code{\link{setChIPpeakAnnoGlobals}} to set all global options at once.
}
\keyword{internal}
