% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-prioritization.R
\name{.prioritizeAnnotations}
\alias{.prioritizeAnnotations}
\title{Unified prioritization function}
\usage{
.prioritizeAnnotations(
  annotated_peaks,
  factor_type,
  sequencing_method = NULL,
  ...
)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}

\item{factor_type}{Factor type}

\item{sequencing_method}{Sequencing method (for PolII, RBP)}

\item{...}{Additional parameters}
}
\value{
GRanges with priority_score column
}
\description{
Unified prioritization function
}
\keyword{internal}
