% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-prioritization.R
\name{.prioritizeChromatinRemodelerAnnotations}
\alias{.prioritizeChromatinRemodelerAnnotations}
\title{Prioritize chromatin remodeler annotations}
\usage{
.prioritizeChromatinRemodelerAnnotations(annotated_peaks)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize chromatin remodeler annotations
}
\keyword{internal}
