% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-prioritization.R
\name{.prioritizeEnhancerHistoneAnnotations}
\alias{.prioritizeEnhancerHistoneAnnotations}
\title{Prioritize enhancer histone mark annotations}
\usage{
.prioritizeEnhancerHistoneAnnotations(
  annotated_peaks,
  mark_type = c("H3K27ac", "H3K4me1"),
  max_enhancer_distance = 5000L
)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}

\item{mark_type}{"H3K27ac" or "H3K4me1"}

\item{max_enhancer_distance}{Maximum distance for enhancer scoring}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize enhancer histone mark annotations
}
\keyword{internal}
