% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-prioritization.R
\name{.prioritizeExonSpecificAnnotations}
\alias{.prioritizeExonSpecificAnnotations}
\title{Prioritize exon-specific annotations}
\usage{
.prioritizeExonSpecificAnnotations(annotated_peaks)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize exon-specific annotations
}
\keyword{internal}
