% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-prioritization.R
\name{.prioritizeGeneBodyHistoneAnnotations}
\alias{.prioritizeGeneBodyHistoneAnnotations}
\title{Prioritize gene body histone mark annotations}
\usage{
.prioritizeGeneBodyHistoneAnnotations(
  annotated_peaks,
  mark_type = c("H3K36me3", "H3K27me3")
)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}

\item{mark_type}{"H3K36me3" or "H3K27me3"}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize gene body histone mark annotations
}
\keyword{internal}
