% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-prioritization.R
\name{.prioritizeIntergenicAnnotations}
\alias{.prioritizeIntergenicAnnotations}
\title{Prioritize intergenic annotations}
\usage{
.prioritizeIntergenicAnnotations(
  annotated_peaks,
  max_intergenic_distance = 50000L
)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}

\item{max_intergenic_distance}{Maximum distance for intergenic scoring}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize intergenic annotations
}
\keyword{internal}
