% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-prioritization.R
\name{.prioritizeIntronSpecificAnnotations}
\alias{.prioritizeIntronSpecificAnnotations}
\title{Prioritize intron-specific annotations}
\usage{
.prioritizeIntronSpecificAnnotations(annotated_peaks)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize intron-specific annotations
}
\keyword{internal}
