% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-prioritization.R
\name{.prioritizeMultiModalAnnotations}
\alias{.prioritizeMultiModalAnnotations}
\title{Prioritize multimodal annotations}
\usage{
.prioritizeMultiModalAnnotations(annotated_peaks, ...)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}

\item{...}{Additional parameters}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize multimodal annotations
}
\keyword{internal}
