% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-prioritization.R
\name{.prioritizeNonCodingRNAAnnotations}
\alias{.prioritizeNonCodingRNAAnnotations}
\title{Prioritize non-coding RNA annotations}
\usage{
.prioritizeNonCodingRNAAnnotations(annotated_peaks)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize non-coding RNA annotations
}
\keyword{internal}
