% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-prioritization.R
\name{.prioritizeRepetitiveElementAnnotations}
\alias{.prioritizeRepetitiveElementAnnotations}
\title{Prioritize repetitive element annotations}
\usage{
.prioritizeRepetitiveElementAnnotations(annotated_peaks)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize repetitive element annotations
}
\keyword{internal}
