% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-prioritization.R
\name{.prioritizeTFAnnotations}
\alias{.prioritizeTFAnnotations}
\title{Prioritize TF annotations}
\usage{
.prioritizeTFAnnotations(annotated_peaks, max_promoter_distance = 2000L)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}

\item{max_promoter_distance}{Maximum distance for promoter scoring}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize TF annotations
}
\keyword{internal}
