% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-prioritization.R
\name{.prioritizeThreeEndAnnotations}
\alias{.prioritizeThreeEndAnnotations}
\title{Prioritize 3' end annotations}
\usage{
.prioritizeThreeEndAnnotations(annotated_peaks, max_tes_distance = 3000L)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}

\item{max_tes_distance}{Maximum distance to TES}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize 3' end annotations
}
\keyword{internal}
