% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-annotatePeaksWithcCRE.R
\name{downloadAllcCREs}
\alias{downloadAllcCREs}
\title{Download all cCREs for a species}
\usage{
downloadAllcCREs(species, cache_dir)
}
\arguments{
\item{species}{Character string, species name}

\item{cache_dir}{Character string, cache directory}
}
\value{
GRanges object with cCRE data
}
\description{
Download all cCREs for a species
}
\keyword{internal}
