% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-annotatePeaksWithcCRE.R
\name{downloadcCREFromURL}
\alias{downloadcCREFromURL}
\title{Download a cCRE file from URL to cache}
\usage{
downloadcCREFromURL(url, cache_file)
}
\arguments{
\item{url}{Character string, URL to download from}

\item{cache_file}{Character string, path to cache file}
}
\value{
GRanges object with cCRE data, or empty GRanges on failure
}
\description{
Download a cCRE file from URL to cache
}
\keyword{internal}
