% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-annotatePeaksWithRoadmap.R
\name{downloadRoadmapData}
\alias{downloadRoadmapData}
\title{Download Roadmap data}
\usage{
downloadRoadmapData(data_type, EIDs, tissue_types, cache_dir, combine_regions)
}
\arguments{
\item{data_type}{Character string, data type}

\item{EIDs}{Character vector, epigenome IDs}

\item{tissue_types}{Character vector, tissue types}

\item{cache_dir}{Character string, cache directory}

\item{combine_regions}{Logical, whether to combine regions (currently always TRUE)}
}
\value{
GRanges object with Roadmap data (always combined)
}
\description{
Download Roadmap data
}
\keyword{internal}
