% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-annotatePeaksWithcCRE.R
\name{downloadTissueSpecificcCREs}
\alias{downloadTissueSpecificcCREs}
\title{Download tissue-specific cCREs}
\usage{
downloadTissueSpecificcCREs(species, tissue_types, cache_dir)
}
\arguments{
\item{species}{Character string, species name}

\item{tissue_types}{Character vector, tissue type names}

\item{cache_dir}{Character string, cache directory}
}
\value{
GRanges object with cCRE data
}
\description{
Download tissue-specific cCREs
}
\keyword{internal}
