% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawVenn.R
\name{DrawVenn}
\alias{DrawVenn}
\title{DrawVenn}
\usage{
DrawVenn(out.dir.name)
}
\arguments{
\item{VennDiagram}{}
}
\description{
DrawVenn
}
\examples{

out.dir.name="/media/H_driver/2016/Yang/Results/"
DrawVenn(out.dir.name)

}

