% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-data-prebuilt.annodata.R
\docType{data}
\name{ExonPlusUtr.human.GRCh37}
\alias{ExonPlusUtr.human.GRCh37}
\title{Gene model with exon, 5' UTR and 3' UTR information for the human genome (GRCh37/hg19)}
\format{
A \code{\link[GenomicRanges]{GRanges}} object with the following
  structure:
  \describe{
    \item{\code{seqnames}}{Chromosome names (e.g., "chr1", "chr2")}
    \item{\code{ranges}}{IRanges object with start and end positions of exons}
    \item{\code{strand}}{Strand information: "+" for positive strand, "-" for negative strand}
    \item{\code{names}}{Ensembl transcript IDs}
    \item{\code{description}}{Transcript description from Ensembl}
    \item{\code{ensembl_gene_id}}{Ensembl gene ID associated with the transcript}
    \item{\code{utr5start}}{Start position of the 5' UTR (untranslated region)}
    \item{\code{utr5end}}{End position of the 5' UTR}
    \item{\code{utr3start}}{Start position of the 3' UTR}
    \item{\code{utr3end}}{End position of the 3' UTR}
  }
}
\source{
Data obtained from Ensembl (GRCh37 release) via biomaRt:
\url{https://grch37.ensembl.org/}
}
\usage{
ExonPlusUtr.human.GRCh37
}
\description{
A pre-computed annotation dataset containing exon, 5' UTR, and 3' UTR
information for *Homo sapiens* based on the GRCh37/hg19 genome assembly.
This dataset was obtained from Ensembl via biomaRt and can be used directly
for peak annotation without requiring an active biomaRt connection.
}
\details{
This dataset was generated using the following code:

\preformatted{
library(biomaRt)
mart <- useMart(
    biomart = "ENSEMBL_MART_ENSEMBL",
    host = "grch37.ensembl.org",
    path = "/biomart/martservice",
    dataset = "hsapiens_gene_ensembl"
)
ExonPlusUtr.human.GRCh37 <- getAnnotation(mart = mart, 
                                          featureType = "ExonPlusUtr")
}

**Note**: GRCh37 (also known as hg19) is an older genome assembly. For
current analyses, consider using GRCh38/hg38 annotations when available.
}
\examples{

# Load the dataset
data(ExonPlusUtr.human.GRCh37)

# Inspect the structure
ExonPlusUtr.human.GRCh37

# Check the number of exons
length(ExonPlusUtr.human.GRCh37)

# View metadata columns
mcols(ExonPlusUtr.human.GRCh37)

# Example: Find exons for a specific gene
# (Replace "ENSG00000139618" with an actual Ensembl gene ID)
# gene_exons <- ExonPlusUtr.human.GRCh37[
#     mcols(ExonPlusUtr.human.GRCh37)$ensembl_gene_id == "ENSG00000139618"
# ]

}
\keyword{datasets}
