% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-overlap-findOverlapsOfPeaks.R
\name{findOverlapsOfPeaks}
\alias{findOverlapsOfPeaks}
\alias{overlappingPeaks}
\alias{overlappingPeaks-class}
\title{Find overlapping peaks among two or more peak sets}
\usage{
findOverlapsOfPeaks(
  ...,
  maxgap = -1L,
  minoverlap = 0L,
  ignore.strand = TRUE,
  connectedPeaks = c("keepAll", "min", "merge")
)
}
\arguments{
\item{\dots}{Two to five objects of \link[GenomicRanges:GRanges-class]{GRanges}
containing peak sets to compare. Alternatively, a single list of GRanges
objects can be provided. Peak sets must have unique names (auto-generated
if missing). See examples below.}

\item{maxgap}{An integer specifying the maximum gap (in base pairs) allowed
between ranges for them to be considered overlapping. Default is \code{-1L},
which means ranges must actually overlap (no gap allowed). See
\code{\link[IRanges:findOverlaps-methods]{findOverlaps}} for details.}

\item{minoverlap}{An integer or numeric value specifying the minimum overlap
required. If an integer >= 1, it specifies the minimum number of base pairs
that must overlap. If \code{0 < minoverlap < 1}, it specifies the minimum
percentage of the interval that must be covered, and the filter condition
will be set to the maximum covered percentage of overlapping peaks.
Default is \code{0L} (any overlap is considered).}

\item{ignore.strand}{A logical value. When \code{TRUE} (default), strand
information is ignored in overlap calculations. When \code{FALSE}, only
ranges on the same strand are considered for overlap.}

\item{connectedPeaks}{A character string specifying how to count connected/overlapping
peak groups. Options:
\itemize{
  \item \code{"keepAll"} (default): Adds the number of involved peaks for
        each peak list to the corresponding overlapping counts. Also outputs
        counts as if \code{connectedPeaks} were set to "min". For example,
        if 5 peaks in group1 overlap with 2 peaks in group2, this will add
        5 peaks to count.group1, 2 to count.group2, and 2 to counts.
  \item \code{"min"}: Adds the minimal number of involved peaks in each group
        of connected/overlapped peaks to the overlapping counts. In the
        example above, this would add 2 to the overlapping counts.
  \item \code{"merge"}: Counts each group of connected peaks as only 1,
        regardless of how many peaks are involved. In the example above, this
        would add 1 to the overlapping counts.
}
See \url{https://support.bioconductor.org/p/133486/#133603} for more examples.}
}
\value{
Returns an object of class \code{overlappingPeaks} containing:
       \itemize{
         \item \code{venn_cnt}: An object of class \code{VennCounts} containing
               the overlap counts for Venn diagram generation
         \item \code{peaklist}: A named list of \code{GRanges} objects, where each
               element represents peaks in a specific overlap category (e.g., peaks
               unique to one set, peaks shared by two sets, etc.). Names are
               constructed from the peak set names joined by "///"
         \item \code{uniquePeaks}: A \code{GRanges} object containing all peaks
               that are unique to a single peak set (not overlapping with any
               other set)
         \item \code{mergedPeaks}: A \code{GRanges} object containing all merged
               overlapping peaks (peaks that overlap between two or more sets)
         \item \code{peaksInMergedPeaks}: A \code{GRanges} object containing all
               original peaks from each sample that are involved in overlapping
               peaks
         \item \code{overlappingPeaks}: A named list of data frames, where each
               data frame contains detailed annotation of overlapping peaks between
               two specific peak sets. Each data frame includes:
               \itemize{
                 \item \code{peaks1}, \code{peaks2}: Peak names from each set
                 \item Genomic coordinates (seqnames, start, end, strand, width)
                 \item \code{overlapFeature}: Spatial relationship between peaks
                       ("upstream", "downstream", "inside", "overlapStart",
                       "overlapEnd", "includeFeature", "overlap")
                 \item \code{shortestDistance}: Shortest distance between the
                       overlapping peaks
               }
         \item \code{all.peaks}: A list of \code{GRanges} objects containing the
               input peaks with formatted names (peak set name + "__" + original
               peak name)
       }
}
\description{
Identifies overlapping genomic regions between two to five sets of peaks using
an efficient interval tree algorithm implemented in GenomicRanges. This function
is the recommended approach for finding overlaps between multiple peak sets,
replacing the deprecated \code{\link{findOverlappingPeaks}} function.

The function efficiently performs overlap queries and provides detailed
information about spatial relationships between overlapping peaks, including
distance calculations and overlap classifications. It returns an object of class
\code{overlappingPeaks} containing Venn counts, merged peaks, unique peaks, and
detailed overlap annotations.
}
\examples{

peaks1 <- GRanges(seqnames=c(6,6,6,6,5),
                 IRanges(start=c(1543200,1557200,1563000,1569800,167889600),
                         end=c(1555199,1560599,1565199,1573799,167893599),
                         names=c("p1","p2","p3","p4","p5")),
                 strand="+")
peaks2 <- GRanges(seqnames=c(6,6,6,6,5),
                  IRanges(start=c(1549800,1554400,1565000,1569400,167888600),
                          end=c(1550599,1560799,1565399,1571199,167888999),
                          names=c("f1","f2","f3","f4","f5")),
                  strand="+")
t1 <- findOverlapsOfPeaks(peaks1, peaks2, maxgap=1000)
makeVennDiagram(t1)
t1$venn_cnt
t1$peaklist
t2 <- findOverlapsOfPeaks(peaks1, peaks2, minoverlap = .5)
makeVennDiagram(t2)

t3 <- findOverlapsOfPeaks(peaks1, peaks2, minoverlap = .90)
makeVennDiagram(t3)

}
\references{
1.Interval tree algorithm from: Cormen, Thomas H.; Leiserson,
Charles E.; Rivest, Ronald L.; Stein, Clifford. Introduction to Algorithms,
second edition, MIT Press and McGraw-Hill. ISBN 0-262-53196-8

2.Zhu L.J. et al. (2010) ChIPpeakAnno: a Bioconductor package to annotate
ChIP-seq and ChIP-chip data. BMC Bioinformatics 2010,
11:237doi:10.1186/1471-2105-11-237

3. Zhu L (2013). "Integrative analysis of ChIP-chip and ChIP-seq dataset."
In Lee T and Luk ACS (eds.), Tilling Arrays, volume 1067, chapter 4, pp.
-19. Humana Press. http://dx.doi.org/10.1007/978-1-62703-607-8_8,
http://link.springer.com/protocol/10.1007\%2F978-1-62703-607-8_8
}
\seealso{
\link{annotatePeakInBatch}, \link{makeVennDiagram},
\link{getVennCounts}, \link{findOverlappingPeaks}
}
\author{
Jianhong Ou
}
\keyword{misc}
