% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utilities-globals.R
\name{getChIPpeakAnnoEnsDb}
\alias{getChIPpeakAnnoEnsDb}
\title{Get global EnsDb for ChIPpeakAnno functions}
\usage{
getChIPpeakAnnoEnsDb()
}
\value{
An object of class \code{EnsDb}, or NULL if not set
}
\description{
Retrieves the currently set global EnsDb option for peaks annotation.
}
\examples{
# Get current global EnsDb (may be NULL)
getChIPpeakAnnoEnsDb()

\dontrun{
# Set and retrieve
library(EnsDb.Hsapiens.v86)
setChIPpeakAnnoEnsDb(EnsDb.Hsapiens.v86)
getChIPpeakAnnoEnsDb()
}
}
\keyword{internal}
