% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utilities-globals.R
\name{getChIPpeakAnnoGenome}
\alias{getChIPpeakAnnoGenome}
\title{Get global genome assembly for ChIPpeakAnno functions}
\usage{
getChIPpeakAnnoGenome()
}
\value{
Character string with the genome assembly, or NULL if not set
}
\description{
Retrieves the currently set global genome assembly option for peaks annotation.
}
\examples{
# Get current global genome (may be NULL)
getChIPpeakAnnoGenome()

# Set and retrieve
setChIPpeakAnnoGenome("hg38")
getChIPpeakAnnoGenome()
}
\keyword{internal}
