% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utilities-globals.R
\name{getChIPpeakAnnoSpecies}
\alias{getChIPpeakAnnoSpecies}
\title{Get global species for ChIPpeakAnno functions}
\usage{
getChIPpeakAnnoSpecies()
}
\value{
Character string with the species, or NULL if not set
}
\description{
Retrieves the currently set global species option for peaks annotation.
}
\examples{
# Get current global species (may be NULL)
getChIPpeakAnnoSpecies()

# Set and retrieve
setChIPpeakAnnoSpecies("Homo sapiens")
getChIPpeakAnnoSpecies()
}
\keyword{internal}
