% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utilities-globals.R
\name{getChIPpeakAnnoTxDb}
\alias{getChIPpeakAnnoTxDb}
\title{Get global TxDb for ChIPpeakAnno functions}
\usage{
getChIPpeakAnnoTxDb()
}
\value{
An object of class \code{TxDb} or \code{EnsDb}, or NULL if not set
}
\description{
Retrieves the currently set global TxDb (or EnsDb) option for peaks annotation.
}
\examples{
# Get current global TxDb (may be NULL)
getChIPpeakAnnoTxDb()

\dontrun{
# Set and retrieve
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
setChIPpeakAnnoTxDb(TxDb.Hsapiens.UCSC.hg38.knownGene)
getChIPpeakAnnoTxDb()
}
}
\keyword{internal}
