% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipSeq.R
\name{getGeneBedName}
\alias{getGeneBedName}
\title{ChipSeq:::getGeneBedName("/Volumes/Bioinformatics$/2017/DannyNewData/NewRe2Danny","peaks.csv","/Volumes/Bioinformatics$/2017/DannyNewData/NewRe2Danny")}
\usage{
getGeneBedName(input.file.dir, file.pattern, output.file.dir)
}
\description{
ChipSeq:::getGeneBedName("/Volumes/Bioinformatics$/2017/DannyNewData/NewRe2Danny","peaks.csv","/Volumes/Bioinformatics$/2017/DannyNewData/NewRe2Danny")
}
