% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-annotatePeaksWithRoadmap.R
\name{getLiftOverChain}
\alias{getLiftOverChain}
\title{Get liftOver chain file for genome conversion}
\usage{
getLiftOverChain(from_genome, to_genome, cache_dir)
}
\arguments{
\item{from_genome}{Character string, source genome (e.g., "hg19")}

\item{to_genome}{Character string, target genome (e.g., "hg38")}

\item{cache_dir}{Character string, directory to cache chain files}
}
\value{
Chain object from rtracklayer, or NULL if not available
}
\description{
Get liftOver chain file for genome conversion
}
\keyword{internal}
