% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipSeq.R
\name{GetResultsFromDiffBind}
\alias{GetResultsFromDiffBind}
\title{GetResultsFromDiffBind
 
 Use DiffBind to process peak profiles}
\usage{
GetResultsFromDiffBind(mcf7, Mergereplicates = c("yes", "no"),
  output.file.dir)
}
\arguments{
\item{output.file.dir}{}
}
\description{
GetResultsFromDiffBind
 
 Use DiffBind to process peak profiles
}
\examples{
output.file.dir="/Volumes/Bioinformatics$/2017/DannyNewData/BindDiff"
mcf7=resmcf
re<-GetResultsFromDiffBind(mcf7,"yes",output.file.dir)

}
