% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexBamFile.R
\name{IndexBamFile}
\alias{IndexBamFile}
\title{IndexBamFile}
\usage{
IndexBamFile(input.file.dir, input.file.pattern, index.file, output.file.dir,
  genome)
}
\arguments{
\item{input.file.dir}{}

\item{input.file.pattern}{}

\item{index.file}{}

\item{output.file.dir}{}

\item{genome}{}
}
\description{
IndexBamFile
}

