% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-annotatePeaksWithcCRE.R
\name{listAvailablecCREs}
\alias{listAvailablecCREs}
\title{List available ENCODE cCRE tissue types}
\usage{
listAvailablecCREs(species = NULL)
}
\arguments{
\item{species}{Character string specifying the species. Must be one of:
\itemize{
  \item \code{"Homo sapiens"}: Human cCREs
  \item \code{"Mus musculus"}: Mouse cCREs
}
If \code{NULL} (default), uses the global species option set by
\code{\link{setChIPpeakAnnoSpecies}}. If no global option is set,
an error is raised prompting the user to specify species.}
}
\value{
A data.frame with columns:
       \itemize{
         \item \code{tissue_type}: Character, tissue type identifier
         \item \code{description}: Character, description of the tissue type
         \item \code{biosample_term}: Character, ENCODE biosample term (if available)
         \item \code{file_accession}: Character, ENCODE file accession (if available)
       }
}
\description{
Lists available tissue types for ENCODE cCRE data. This helps users 
discover valid values for the \code{tissue_types} parameter in 
\code{\link{annotatePeaksWithcCRE}}.
}
\details{
This function queries ENCODE/SCREEN to discover available tissue-specific 
cCRE datasets. The returned data.frame can be used to select appropriate 
tissue types for \code{annotatePeaksWithcCRE}.
}
\note{
\itemize{
  \item Requires internet connection to query ENCODE API
  \item Results may vary as ENCODE data is updated
  \item If API query fails, returns a basic data.frame with common tissue types
}
}
\examples{

# List available human cCRE tissue types
\dontrun{
available <- listAvailablecCREs("Homo sapiens")
head(available)

# Use a specific tissue type
data("myPeakList")
# original genome assembly is hg18, for now we set the global genome 
# assembly to hg38 to save time.
setChIPpeakAnnoGlobals(genome = "hg38", species = "Homo sapiens")
annotated <- annotatePeaksWithcCRE(myPeakList,
                                    tissue_types = available$tissue_type[1])
}
}
\seealso{
\code{\link{annotatePeaksWithcCRE}} for using tissue types in annotation
}
\author{
Haibo Liu
}
\keyword{misc}
