% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-annotatePeaksWithcCRE.R
\name{loadcCREsFromFile}
\alias{loadcCREsFromFile}
\title{Load cCRE data from a file or URL}
\usage{
loadcCREsFromFile(file_path)
}
\arguments{
\item{file_path}{Character string, path to cCRE file or URL}
}
\value{
GRanges object with cCRE data
}
\description{
Loads cCRE data from a file or URL. Handles format variations between
all-in-one cCRE files and tissue-specific files, as SCREEN files do not
strictly follow standard BED format.
}
\keyword{internal}
