% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-annotatePeaksWithRoadmap.R
\name{loadRoadmapFromFile}
\alias{loadRoadmapFromFile}
\title{Load Roadmap data from a file or URL}
\usage{
loadRoadmapFromFile(file_path)
}
\arguments{
\item{file_path}{Character string, path to BED file or URL}
}
\value{
GRanges object with Roadmap data
}
\description{
Load Roadmap data from a file or URL
}
\keyword{internal}
