% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergePeak.R
\name{MergePeak}
\alias{MergePeak}
\title{MergePeak}
\usage{
MergePeak(input.file.dir, output.file.dir, genome)
}
\arguments{
\item{input.file.dir}{}

\item{output.file.dir}{}

\item{genome}{}
}
\description{
Merge peaks between replicates for same cell lines
}
\examples{

input.file.dir="/media/H_driver/2016/Danny/Danny_chip_PeakCall/"
output.file.dir="/media/H_driver/2016/Danny/Danny_chip_PeakCall/"

AnnotatePeak3(input.file.dir,output.file.dir,genome="Mm")
AnnotatePeak3(input.file.dir,output.file.dir,genome="Hs")


}

