% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParserBedFile.R
\name{ParserBedFile}
\alias{ParserBedFile}
\title{Title}
\usage{
ParserBedFile(dir.name, input.file.pattern, out.dir.name, output.test.file)
}
\arguments{
\item{input.file.pattern}{}
}
\description{
Title
}
\examples{

dir.name="/media/H_driver/2016/Yang/BedFromPeakCall/"
input.file.pattern="*bam_mm_summits_from_PeakCall4Yang.bed"
out.dir.name="/media/H_driver/2016/Yang/Results/"
output.test.file= "SMC1A_RAD21_overlap_test"

ParserBedFile(dir.name,input.file.pattern,out.dir.name,output.test.file)

}

