% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParserBedFile4Peng.R
\name{ParserBedFile4Peng}
\alias{ParserBedFile4Peng}
\title{ParserBedFile4Peng}
\usage{
ParserBedFile4Peng(dir.name, input.file.pattern, out.dir.name)
}
\arguments{
\item{input.file.pattern}{}
}
\description{
ParserBedFile4Peng
}
\examples{

dir.name="/media/H_driver/2016/Yang/BedFromPeng/"
input.file.pattern="*.bed"

out.dir.name="/media/H_driver/2016/Yang/Results/"

re.from.bed.peng<-ParserBedFile4Peng(dir.name,input.file.pattern,out.dir.name)


save.image(file=paste0(out.dir.name,"re_save_peng.RData"))
savehistory(file=paste0(out.dir.name,"re_save_peng.Rhistory"))

}

