% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParserBedFile4PengDiffBind.R
\name{ParserBedFile4PengDiffBind}
\alias{ParserBedFile4PengDiffBind}
\title{ParserBedFile4PengDiffBind}
\usage{
ParserBedFile4PengDiffBind(dir.name, input.file.pattern, out.dir.name)
}
\arguments{
\item{input.file.pattern}{}
}
\description{
ParserBedFile4PengDiffBind
}
\examples{

dir.name="/media/H_driver/2016/Yang/BedFromPeng/"
input.file.pattern="*peaks_from_PeakCall4Yang.bed"
out.dir.name="/media/H_driver/2016/Yang/Results/"

re.from.bed.peng.4.venn<-ParserBedFile4PengVenn(dir.name,input.file.pattern,out.dir.name)


save.image(file=paste0(out.dir.name,"re_save_peng.RData"))
savehistory(file=paste0(out.dir.name,"re_save_peng.Rhistory"))

}

