% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipSeq.R
\name{ParserReadFiles}
\alias{ParserReadFiles}
\title{ParserReadFiles}
\usage{
ParserReadFiles(input.file.dir, input.file.type)
}
\arguments{
\item{input.file.dir}{}

\item{output.file.dir}{}
}
\description{
ParserReadFiles
}
\examples{

input.file.dir="~/TestBam/"
out.file.dir="~/"
testR<-ParserReadFiles(input.file.dir) 


}
