% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakCallAndAnnotation.R
\name{PeakCallAndAnnotation}
\alias{PeakCallAndAnnotation}
\title{PeakCallAndAnnotation}
\usage{
PeakCallAndAnnotation(input.file.dir, output.file.dir, genome)
}
\arguments{
\item{input.file.dir}{}

\item{output.file.dir}{}

\item{genome}{}

\item{input.file.pattern}{}

\item{index.file}{}
}
\description{
Call peak for bam files using macs14 and perform peak annotation using ChIPpeakAnno
}
