% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-data-peaksets.R
\docType{data}
\name{Peaks.Ste12.Replicate1}
\alias{Peaks.Ste12.Replicate1}
\title{Ste12-binding sites from biological replicate 1 in yeasts}
\format{
A \code{\link[GenomicRanges]{GRanges}} object with the following structure:
\describe{
  \item{seqnames}{Chromosome names (yeast chromosomes: "chrI", "chrII", etc.)}
  \item{ranges}{\code{\link[IRanges]{IRanges}} object with start and end positions}
  \item{strand}{Strand information ("+", "-", or "*")}
  \item{names}{Peak identifiers as character vector}
}
}
\usage{
Peaks.Ste12.Replicate1
}
\description{
A \code{\link[GenomicRanges]{GRanges}} object containing Ste12 transcription
factor binding sites identified from biological replicate 1 in yeast
(\emph{Saccharomyces cerevisiae}) ChIP-seq experiments. This dataset is part
of a three-replicate series used to demonstrate overlap analysis and
reproducibility assessment.
}
\details{
Ste12 is a transcription factor involved in mating and filamentous growth in
yeast. These datasets contain binding sites identified from three biological
replicates, which can be used to:
\itemize{
  \item Demonstrate overlap analysis between replicates
  \item Assess reproducibility using \code{\link{findOverlapsOfPeaks}}
  \item Generate Venn diagrams with \code{\link{makeVennDiagram}}
  \item Perform permutation testing with \code{\link{peakPermTest}}
}
}
\examples{
# Load replicate 1
data(Peaks.Ste12.Replicate1)

# Inspect the data
Peaks.Ste12.Replicate1
length(Peaks.Ste12.Replicate1)

# Compare with other replicates
data(Peaks.Ste12.Replicate2)
data(Peaks.Ste12.Replicate3)

# Find overlaps between replicates
overlaps <- findOverlapsOfPeaks(
    Peaks.Ste12.Replicate1,
    Peaks.Ste12.Replicate2,
    Peaks.Ste12.Replicate3
)
}
\references{
Lefranois P, Euskirchen GM, Auerbach RK, Rozowsky J, Gibson T, Yellman CM,
Gerstein M, Snyder M (2009) Efficient yeast ChIP-Seq using multiplex
short-read DNA sequencing. \emph{BMC Genomics} 10:37.
\doi{10.1186/1471-2164-10-37}
}
\seealso{
\code{\link{Peaks.Ste12.Replicate2}}, \code{\link{Peaks.Ste12.Replicate3}},
\code{\link{findOverlapsOfPeaks}}, \code{\link{makeVennDiagram}}
}
\keyword{datasets}
