% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFileSortIndexVisualization.R
\name{plotBam}
\alias{plotBam}
\title{R -e 'library(ChipSeq);ChipSeq:::plotBam(input.file.dir="/scratch/projects/bbc/Project/Danny_chip2/Alignment/BWA",file.type="*marked.bam",output.file.dir="/scratch/projects/bbc/aiminy_project/DannyNewNgsPlot")'}
\usage{
plotBam(input.file.dir, file.type, output.file.dir, BigMem = FALSE,
  cores = 15, Memory = 25000, Wall.time = "72:00", span.ptile = 8)
}
\description{
R -e 'library(ChipSeq);ChipSeq:::plotBam(input.file.dir="/scratch/projects/bbc/Project/Danny_chip2/Alignment/BWA",file.type="*marked.bam",output.file.dir="/scratch/projects/bbc/aiminy_project/DannyNewNgsPlot")'
}
