% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-prioritizeAnnotation.R
\name{prioritizeAnnotations}
\alias{prioritizeAnnotations}
\title{Unified prioritization function}
\usage{
prioritizeAnnotations(
  annotated_peaks,
  factor_type,
  sequencing_method = c("ChIP-seq", "GRO-seq", "PRO-seq", "CLIP-seq", "iCLIP", "eCLIP")
)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}

\item{factor_type}{Factor type}

\item{sequencing_method}{Sequencing method. For PolII, it can be "ChIP-seq", "GRO-seq", "PRO-seq". 
For RBP, it can be "ChIP-seq", "CLIP-seq", "iCLIP", "eCLIP".}
}
\value{
GRanges with priority_score column
}
\description{
Unified prioritization function
}
