% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-prioritizeAnnotation.R
\name{prioritizeArchitecturalAnnotations}
\alias{prioritizeArchitecturalAnnotations}
\title{Prioritize architectural feature annotations}
\usage{
prioritizeArchitecturalAnnotations(annotated_peaks)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize architectural feature annotations
}
