% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-prioritizeAnnotation.R
\name{prioritizePolIIAnnotations}
\alias{prioritizePolIIAnnotations}
\title{Prioritize Pol II annotations}
\usage{
prioritizePolIIAnnotations(
  annotated_peaks,
  sequencing_method = c("ChIP-seq", "GRO-seq", "PRO-seq")
)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}

\item{sequencing_method}{"ChIP-seq", "GRO-seq", or "PRO-seq"}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize Pol II annotations
}
