% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-prioritizeAnnotation.R
\name{prioritizePromoterHistoneAnnotations}
\alias{prioritizePromoterHistoneAnnotations}
\title{Prioritize promoter histone mark annotations}
\usage{
prioritizePromoterHistoneAnnotations(
  annotated_peaks,
  mark_type = c("H3K4me3", "H3K4me2"),
  max_promoter_distance = 3000L
)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}

\item{mark_type}{"H3K4me3" or "H3K4me2"}

\item{max_promoter_distance}{Maximum distance for promoter scoring}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize promoter histone mark annotations
}
