% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-annotation-prioritizeAnnotation.R
\name{prioritizeRBPAnnotations}
\alias{prioritizeRBPAnnotations}
\title{Prioritize RBP annotations}
\usage{
prioritizeRBPAnnotations(
  annotated_peaks,
  sequencing_method = c("ChIP-seq", "CLIP-seq", "iCLIP", "eCLIP"),
  use_transcript_level = TRUE
)
}
\arguments{
\item{annotated_peaks}{GRanges with annotations}

\item{sequencing_method}{"ChIP-seq", "CLIP-seq", "iCLIP", or "eCLIP"}

\item{use_transcript_level}{Logical, prefer transcript-level annotation}
}
\value{
GRanges with priority_score column
}
\description{
Prioritize RBP annotations
}
