% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessUsingCHIPpeakAnno.R
\name{ProcessUsingCHIPpeakAnno}
\alias{ProcessUsingCHIPpeakAnno}
\title{ProcessUsingCHIPpeakAnno}
\usage{
ProcessUsingCHIPpeakAnno(dir.name, input.file.pattern, out.dir.name)
}
\arguments{
\item{out.dir.name}{}
}
\description{
ProcessUsingCHIPpeakAnno
}
\examples{
 dir.name="/media/H_driver/2016/Yang/MACS/MACS/"
 input.file.pattern="*.bed"
 out.dir.name="/media/H_driver/2016/Yang/Results/"


}

