% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessUsingChipSeeker.R
\name{ProcessUsingChipSeeker}
\alias{ProcessUsingChipSeeker}
\title{Title}
\usage{
ProcessUsingChipSeeker(which.bed.file)
}
\description{
Title
}
\examples{

which.bed.file="/media/H_driver/2016/Yang/Results/WT_triple_overlap.bed"

}

