% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipSeq.R
\name{reformatPath}
\alias{reformatPath}
\title{Title}
\usage{
reformatPath(dir.name)
}
\arguments{
\item{dir.name}{}
}
\description{
Title
}
\examples{
dir.name="/media/H_driver/2016/Yang/MACS/MACS/"
reformatPath(dir.name) 
}
