% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipSeq.R
\name{runPlotBam2}
\alias{runPlotBam2}
\title{R -e 'library(DoGs);library(ChipSeq);ChipSeq:::runPlotBam2("~/Danny/SampleID_INFO_ChIP_new_Danny.csv","~/Danny/Alignment/BWA/bam_files.txt","/scratch/projects/bbc/aiminy_project/DannyBamMerged")'}
\usage{
runPlotBam2(input.sample.file, input.bam.file, output.dir)
}
\description{
R -e 'library(DoGs);library(ChipSeq);ChipSeq:::runPlotBam2("~/Danny/SampleID_INFO_ChIP_new_Danny.csv","~/Danny/Alignment/BWA/bam_files.txt","/scratch/projects/bbc/aiminy_project/DannyBamMerged")'
}
