% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utilities-globals.R
\name{setChIPpeakAnnoEnsDb}
\alias{setChIPpeakAnnoEnsDb}
\title{Set global EnsDb for ChIPpeakAnno functions}
\usage{
setChIPpeakAnnoEnsDb(EnsDb = NULL)
}
\arguments{
\item{EnsDb}{An object of class \code{\link[ensembldb:EnsDb-class]{EnsDb}}
containing genomic annotation. If NULL, clears the global EnsDb option.}
}
\value{
Invisibly returns the previous value of the global EnsDb option
}
\description{
Sets a global default EnsDb object that will be used by ChIPpeakAnno
functions when an EnsDb parameter is not explicitly provided. This provides
a convenient way to set defaults that are used across multiple functions, while
still allowing per-function overrides.
}
\details{
The global EnsDb option is used as a fallback when:
\itemize{
  \item Functions require an EnsDb parameter but it's not provided
  \item Functions have \code{EnsDb = NULL} as default
}

The global option can always be overridden by explicitly providing EnsDb
in individual function calls.
}
\examples{
\dontrun{
library(EnsDb.Hsapiens.v86)

# Set global EnsDb
setChIPpeakAnnoEnsDb(EnsDb.Hsapiens.v86)

# Get the current global EnsDb
getChIPpeakAnnoEnsDb()

# Clear the global EnsDb (set to NULL)
setChIPpeakAnnoEnsDb(NULL)
}
}
\keyword{internal}
