% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utilities-globals.R
\name{setChIPpeakAnnoSpecies}
\alias{setChIPpeakAnnoSpecies}
\title{Set global species for ChIPpeakAnno functions}
\usage{
setChIPpeakAnnoSpecies(species = NULL)
}
\arguments{
\item{species}{Character string specifying the species (e.g., "Homo sapiens",
"Mus musculus"). Should match the species names used in ChIPpeakAnno
functions.}
}
\value{
Invisibly returns the previous value of the global species option
}
\description{
Sets a global default species that will be used by ChIPpeakAnno functions
when species is not explicitly provided. This provides convenience for
users who consistently work with one species.
}
\details{
The global species option is used as a default when:
\itemize{
  \item Species parameter is not provided to a function
  \item Function requires species but user wants to use global default
}

The global option can always be overridden by explicitly setting the species
in individual function calls.
}
\examples{
# Set global species to human
setChIPpeakAnnoSpecies("Homo sapiens")

# Get the current global species
getChIPpeakAnnoSpecies()

# Clear the global species (set to NULL)
setChIPpeakAnnoSpecies(NULL)
}
\keyword{internal}
