% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utilities-globals.R
\name{setChIPpeakAnnoTxDb}
\alias{setChIPpeakAnnoTxDb}
\title{Set global TxDb for ChIPpeakAnno functions}
\usage{
setChIPpeakAnnoTxDb(TxDb = NULL)
}
\arguments{
\item{TxDb}{An object of class \code{\link[GenomicFeatures:TxDb-class]{TxDb}}
or \code{\link[ensembldb:EnsDb-class]{EnsDb}} containing genomic
annotation. If NULL, clears the global TxDb option.}
}
\value{
Invisibly returns the previous value of the global TxDb option
}
\description{
Sets a global default TxDb (or EnsDb) object that will be used by ChIPpeakAnno
functions when a TxDb/EnsDb parameter is not explicitly provided. This provides
a convenient way to set defaults that are used across multiple functions, while
still allowing per-function overrides.
}
\details{
The global TxDb option is used as a fallback when:
\itemize{
  \item Functions require a TxDb/EnsDb parameter but it's not provided
  \item Functions have \code{TxDb = NULL} or \code{EnsDb = NULL} as default
}

The global option can always be overridden by explicitly providing TxDb/EnsDb
in individual function calls.
}
\examples{
\dontrun{
library(TxDb.Hsapiens.UCSC.hg38.knownGene)

# Set global TxDb
setChIPpeakAnnoTxDb(TxDb.Hsapiens.UCSC.hg38.knownGene)

# Get the current global TxDb
getChIPpeakAnnoTxDb()

# Clear the global TxDb (set to NULL)
setChIPpeakAnnoTxDb(NULL)
}
}
\keyword{internal}
