% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortBamFile.R
\name{SortBamFile}
\alias{SortBamFile}
\title{SortBamFile}
\usage{
SortBamFile(input.file.dir, input.file.pattern, index.file, output.file.dir,
  genome)
}
\arguments{
\item{input.file.dir}{}

\item{input.file.pattern}{}

\item{index.file}{}

\item{output.file.dir}{}

\item{genome}{}
}
\description{
SortBamFile
}

