% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipSeq.R
\name{usePegasus}
\alias{usePegasus}
\title{x <- ChipSeq:::usePegasus('parallel', Wall.time = '72:00',cores = 32,Memory = 16000,span.ptile = 16,'bamPlot')}
\usage{
usePegasus(job.option = c("general", "parallel", "bigmem"), Wall.time, cores,
  Memory, span.ptile, job.name, wait.job.name = NULL)
}
\description{
x <- ChipSeq:::usePegasus('parallel', Wall.time = '72:00',cores = 32,Memory = 16000,span.ptile = 16,'bamPlot')
}
