% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamFileSortIndexVisualization.R
\name{BamFileSortIndexVisualization}
\alias{BamFileSortIndexVisualization}
\title{BamFileSortIndexVisualization}
\usage{
BamFileSortIndexVisualization(input.file.dir, output.file.dir, genome)
}
\arguments{
\item{input.file.dir}{}

\item{output.file.dir}{}

\item{genome}{}
}
\description{
BamFileSortIndexVisualization
}
