% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipSeq.R
\name{BamFileSortIndexVisualization}
\alias{BamFileSortIndexVisualization}
\title{BamFileSortIndexVisualization}
\usage{
BamFileSortIndexVisualization(input.file.dir, output.file.dir, genome)
}
\arguments{
\item{input.file.dir}{}

\item{output.file.dir}{}

\item{genome}{}
}
\description{
BamFileSortIndexVisualization
}
\examples{

input.file.dir="/projects/scratch/bbc/Project/Danny_chip/Alignment/BWA/"
output.file.dir="/scratch/projects/bbc/aiminy_project/"
genome="Hs"

BamFileSortIndexVisualization(input.file.dir,output.file.dir,genome)

}
