% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-internal-bdp.R
\name{bdp}
\alias{bdp}
\title{Obtain peaks near bi-directional promoters}
\usage{
bdp(peaks, annoData, maxgap = 2000L, ...)
}
\arguments{
\item{peaks}{A \code{GRanges} object containing peak regions.}

\item{annoData}{An \code{annoGR} or \code{GRanges} object containing
annotation data (e.g., TSS positions).}

\item{maxgap}{Integer. Maximum distance (in base pairs) between a peak and
TSS to be considered "near". Default is 2000.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A \code{GRangesList} object containing peaks near bi-directional
  promoters. Each element in the list corresponds to a peak that is within
  \code{maxgap} distance of promoters on both the plus and minus strands.
  Returns \code{NA} if no peaks meet the criteria.
}
\description{
Identifies peaks that are located near bi-directional promoters (BDPs), which
are genomic regions where two genes are transcribed in opposite directions from
a shared promoter region. A peak is considered near a BDP if it is within
\code{maxgap} distance of promoters on both strands.
}
\examples{

  if(interactive() || Sys.getenv("USER")=="jou"){
    library(ensembldb)
    library(EnsDb.Hsapiens.v75)
    data("myPeakList")
    annoGR <- annoGR(EnsDb.Hsapiens.v75)
    seqlevelsStyle(myPeakList) <- seqlevelsStyle(annoGR)
    ChIPpeakAnno:::bdp(myPeakList, annoGR)
  }

}
\seealso{
See Also as \code{\link{annoPeaks}}, \code{\link{annoGR}}
}
\author{
Jianhong Ou
}
\keyword{misc}
