% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cntOverlaps.R
\name{cntOverlaps}
\alias{cntOverlaps}
\title{count overlaps}
\usage{
cntOverlaps(A, B, maxgap = 0L, ...)
}
\arguments{
\item{A, B}{A GRanges object.}

\item{maxgap}{A single integer >= 0.}

\item{...}{parameters passed to \link[regioneR]{numOverlaps}#'}
}
\description{
Count overlaps with max gap.
}
