% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egOrgMap.R
\name{egOrgMap}
\alias{egOrgMap}
\title{Convert between the name of the organism annotation package ("OrgDb") and
the name of the organism.}
\usage{
egOrgMap(name)
}
\arguments{
\item{name}{The name of the organism annotation package or the species.}
}
\value{
A object of character
}
\description{
Give a species name and return the organism annotation package name or give
an organism annotation package name then return the species name.
}
\examples{

  egOrgMap("org.Hs.eg.db")
  egOrgMap("Mus musculus")

}
\author{
Jianhong Ou
}
\keyword{misc}
