% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privateUtil.R
\name{EnsDb2GR}
\alias{EnsDb2GR}
\title{EnsDb object to GRanges}
\usage{
EnsDb2GR(ranges, feature)
}
\arguments{
\item{ranges}{an EnsDb object}

\item{feature}{feature type, could be disjointExons, gene, exon and 
transcript}
}
\description{
convert EnsDb object to GRanges
}
