% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResultsFromDiffBind.R
\name{GetResultsFromDiffBind}
\alias{GetResultsFromDiffBind}
\title{GetResultsFromDiffBind
 
 Use DiffBind to process peak profiles}
\usage{
GetResultsFromDiffBind(mcf7, output.file.dir)
}
\arguments{
\item{output.file.dir}{}
}
\description{
GetResultsFromDiffBind
 
 Use DiffBind to process peak profiles
}
\examples{

data(mcf7)
output.file.dir="/Volumes/Bioinformatics$/2016/Danny/Analysis4Peaks/"
GetResultsFromDiffBind(mcf7,output.file.dir)


}

