% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergePeak.R
\name{GetSampleInfo}
\alias{GetSampleInfo}
\title{GetSampleInfo}
\usage{
GetSampleInfo(input.file)
}
\arguments{
\item{input.file}{}
}
\description{
GetSampleInfo
}
\examples{

input.file="/Volumes/Bioinformatics$/2016/Danny/SampleID_INFO_ChIP.csv"

re<-GetSampleInfo(input.file)


}

