% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSampleBamFileInfoCallPeak.R
\name{GetSampleInfo}
\alias{GetSampleInfo}
\title{GetSampleInfo}
\usage{
GetSampleInfo(input.sample.file, input.bam.file)
}
\arguments{
\item{input.bam.file}{}
}
\description{
GetSampleInfo
}
\examples{

input.sample.file <- '/Volumes/Bioinformatics$/2017/DannyNewData/SampleID_INFO_ChIP_new_Danny.csv'

input.bam.file <- '/Volumes/Bioinformatics$/2017/DannyNewData/sorted_bam_files.txt'
 
re <- GetSampleInfo(input.sample.file,input.bam.file)


}
