% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniqueGOidCount.R
\name{getUniqueGOidCount}
\alias{getUniqueGOidCount}
\title{get the count for each unique GO ID}
\usage{
getUniqueGOidCount(goList)
}
\arguments{
\item{goList}{a set of GO terms as character vector}
}
\value{
a list with 2 variables \item{GOterm}{a vector of GO terms as
character vector} \item{GOcount}{counts corresponding to the above GOterm as
numeric vector}
}
\description{
get the count for each unique GO ID
}
\note{
internal function not intended to be called directly by users
}
\examples{

goList= c("GO:0000075", "GO:0000082","GO:0000082","GO:0000122","GO:0000122",
           "GO:0000075","GO:0000082","GO:0000082","GO:0000122","GO:0000122",
           "GO:0000122","GO:0000122","GO:0000075", "GO:0000082","GO:000012")
 
getUniqueGOidCount(goList)

}
\seealso{
getEnrichedGO
}
\author{
Lihua Julie Zhu
}
\keyword{internal}
