% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligoSummary.R
\name{oligoFrequency}
\alias{oligoFrequency}
\title{get the oligonucleotide frequency}
\usage{
oligoFrequency(sequence, MarkovOrder = 3L, last = 1e+06)
}
\arguments{
\item{sequence}{The sequences packaged in DNAStringSet, DNAString object or
output of function \link{getAllPeakSequence}.}

\item{MarkovOrder}{Markov order.}

\item{last}{The sequence size to be analyzed.}
}
\value{
A numeric vector.
}
\description{
Prepare the oligonucleotide frequency for given Markov order.
}
\examples{
    library(seqinr)
    library(Biostrings)
    oligoFrequency(DNAString("AATTCGACGTACAGATGACTAGACT"))

}
\seealso{
See Also as \code{\link{oligoSummary}}
}
\author{
Jianhong Ou
}
\keyword{misc}
