% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser4diffbind.R
\name{parser4diffbind}
\alias{parser4diffbind}
\title{parser4diffbind}
\usage{
parser4diffbind(input.sample.file, input.bed.dir, input.file.pattern)
}
\arguments{
\item{input.file.pattern}{}
}
\description{
parser4diffbind
}
\examples{
 
input.sample.file = "/Volumes/Bioinformatics$/2017/DannyNewData/SampleID_INFO_ChIP_new_Danny.csv"
input.bed.dir = "/Volumes/Bioinformatics$/2017/DannyNewData/PeakCall/"
input.file.pattern = "bed"
resmcf<-parser4diffbind(input.sample.file,input.bed.dir,input.file.pattern)

}
