% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParserBedFile.R
\name{ParserBedFile}
\alias{ParserBedFile}
\title{Title}
\usage{
ParserBedFile(dir.name, input.file.pattern, out.dir.name)
}
\arguments{
\item{input.file.pattern}{}
}
\description{
Title
}
\examples{

dir.name="/media/H_driver/2016/Yang/BedFromPeakCall/"
input.file.pattern="*bam_mm_summits_from_PeakCall4Yang.bed"

out.dir.name="/media/H_driver/2016/Yang/Results/"
output.test.file= "SMC1A_RAD21_overlap_test"


dir.name.2="/media/H_driver/2016/Yang/BedFromPeakCall_mm_100e_07/"
input.file.pattern.2="*_bam_mm_1_from_PeakCall4Yang.bed"

out.dir.name.2="/media/H_driver/2016/Yang/Results/"
output.test.file.2= "SMC1A_RAD21_mm_100e-07_peaks_overlap_test"

ParserBedFile(dir.name,input.file.pattern,out.dir.name)

re.mm.100e.07<-ParserBedFile(dir.name.2,input.file.pattern.2,out.dir.name)


save.image(file=paste0(out.dir.name,"re_save_2.RData"))
savehistory(file=paste0(out.dir.name,"re_save_2.Rhistory"))

}
