% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSampleBamFileInfoCallPeak.R
\name{peakcallwithinput}
\alias{peakcallwithinput}
\title{peakcallwithinput}
\usage{
peakcallwithinput(input.sample.file, input.bam.file, genome = c("Hs", "hs",
  "HS", "hS"), output.dir, peakcaller = c("macs14", "macs2"), peakPvalue)
}
\description{
peakcallwithinput
}
