% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peaks.Ste12.R
\docType{data}
\name{Peaks.Ste12.Replicate2}
\alias{Peaks.Ste12.Replicate2}
\title{Ste12-binding sites from biological replicate 2 in yeast (see reference)}
\format{
GRanges with slot names containing the ID of peak as character, slot
start containing the start position of the peak, slot end containing the end
position of the peak and space containing the chromosome where the peak is
located.
}
\source{
http://www.biomedcentral.com/1471-2164/10/37
}
\usage{
Peaks.Ste12.Replicate2
}
\description{
Ste12-binding sites from biological replicate 2 in yeast (see reference)
}
\examples{

data(Peaks.Ste12.Replicate2)
Peaks.Ste12.Replicate2
}
\references{
Philippe Lefranois, Ghia M Euskirchen, Raymond K Auerbach, Joel
Rozowsky, Theodore Gibson, Christopher M Yellman, Mark Gerstein and Michael
Snyder (2009) Efficient yeast ChIP-Seq using multiplex short-read DNA
sequencing BMC Genomics 10:37doi:10.1186/1471-2164-10-37
}
\keyword{datasets}
