\name{Peaks.Ste12.Replicate3}
\alias{Peaks.Ste12.Replicate3}
\docType{data}
\title{
 Ste12-binding sites from biological replicate 3 in yeast (see reference)
}
\description{
 Ste12-binding sites from biological replicate 3 in yeast (see reference)
}
\usage{data(Peaks.Ste12.Replicate3)}
\format{
    RangedData with slot rownames containing the ID of peak as character, slot 
    start containing the start position of the peak, slot end containing the 
    end position of the peak and space containing the chromosome where the 
    peak is located.
}
\source{http://www.biomedcentral.com/1471-2164/10/37
}
\references{
Philippe Lefranois, Ghia M Euskirchen, Raymond K Auerbach, Joel Rozowsky, 
Theodore Gibson, Christopher M Yellman, Mark Gerstein  and Michael Snyder (2009)
Efficient yeast ChIP-Seq using multiplex short-read DNA sequencing 
BMC Genomics 10:37doi:10.1186/1471-2164-10-37
}
\examples{
data(Peaks.Ste12.Replicate3)
Peaks.Ste12.Replicate3
}
\keyword{datasets}
