% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks1.R
\docType{data}
\name{peaks1}
\alias{peaks1}
\title{An example GRanges object representing a ChIP-seq peak dataset}
\format{
GRanges
}
\usage{
peaks1
}
\description{
An example GRanges object representing a ChIP-seq peak dataset
}
\examples{

data(peaks1)
head(peaks1, n = 2)
}
\keyword{datasets}
