% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks2.R
\docType{data}
\name{peaks2}
\alias{peaks2}
\title{An example GRanges object representing a ChIP-seq peak dataset}
\format{
GRanges
}
\usage{
peaks2
}
\description{
An example GRanges object representing a ChIP-seq peak dataset
}
\examples{

data(peaks2)
head(peaks2, n = 2)
}
\keyword{datasets}
