% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks3.R
\docType{data}
\name{peaks3}
\alias{peaks3}
\title{An example GRanges object representing a ChIP-seq peak dataset}
\format{
GRanges
}
\usage{
peaks3
}
\description{
An example GRanges object representing a ChIP-seq peak dataset
}
\examples{

data(peaks3)
head(peaks3, n = 2)
}
\keyword{datasets}
