% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permPool.R
\docType{class}
\name{permPool-class}
\alias{permPool-class}
\alias{permPool}
\alias{permPool-method}
\alias{$,permPool-method}
\alias{$<-,permPool-method}
\title{Class \code{"permPool"}}
\description{
An object of class \code{"permPool"} represents the possible locations to do
permutation test.
}
\section{Slots}{

\describe{
\item{\code{grs}}{object of \code{"GRangesList"} The list of binding ranges}

\item{\code{N}}{vector of \code{"integer"}, permutation number for each ranges}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("permPool", grs="GRangesList", N="integer")}.
}

\seealso{
\link{preparePool}, \link{peakPermTest}
}
\keyword{classes}
