% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessUsingCHIPpeakAnno.R
\name{ProcessUsingCHIPpeakAnno}
\alias{ProcessUsingCHIPpeakAnno}
\title{ProcessUsingCHIPpeakAnno}
\usage{
ProcessUsingCHIPpeakAnno(input.file.dir, input.file.pattern, output.file.dir)
}
\description{
ProcessUsingCHIPpeakAnno
}
\examples{

input.file.dir="/media/H_driver/2016/Yang/MACS/MACS/"
input.file.pattern="*.bed"
output.file.dir="/media/H_driver/2016/Yang/MACS/MACS/"

ProcessUsingCHIPpeakAnno(input.file.dir,input.file.pattern,output.file.dir)

}

