% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reCenterPeaks.R
\name{reCenterPeaks}
\alias{reCenterPeaks}
\title{re-center the peaks}
\usage{
reCenterPeaks(peaks, width = 2000L, ...)
}
\arguments{
\item{peaks}{An object of \link[GenomicRanges:GRanges-class]{GRanges} or
\link{annoGR}.}

\item{width}{The width of new peaks}

\item{...}{Not used.}
}
\value{
An object of GRanges.
}
\description{
Create a new list of peaks based on the peak centers of given list.
}
\examples{

    reCenterPeaks(GRanges("chr1", IRanges(1, 10)), width=2)

}
\author{
Jianhong Ou
}
\keyword{misc}
