% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.human.GRCh38.R
\docType{data}
\name{TSS.human.GRCh38}
\alias{TSS.human.GRCh38}
\title{TSS annotation for human sapiens (GRCh38) obtained from biomaRt}
\format{
A 'GRanges' [package "GenomicRanges"] object with ensembl id as
names.
}
\usage{
TSS.human.GRCh38
}
\description{
TSS annotation for human sapiens (GRCh38) obtained from biomaRt
}
\details{
used in the examples Annotation data obtained by:

mart = useMart(biomart = "ensembl", dataset = "hsapiens_gene_ensembl")

getAnnotation(mart, featureType = "TSS")
}
\examples{

data(TSS.human.GRCh38)
slotNames(TSS.human.GRCh38)

}
\keyword{datasets}
