% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.human.NCBI36.R
\docType{data}
\name{TSS.human.NCBI36}
\alias{TSS.human.NCBI36}
\title{TSS annotation for human sapiens (NCBI36) obtained from biomaRt}
\format{
GRanges with slot start holding the start position of the gene, slot
end holding the end position of the gene, slot names holding ensembl gene
id, slot seqnames holding the chromosome location where the gene is located
and slot strand holding the strinad information. In addition, the following
variables are included.  \describe{ \item{list("description")}{description
of the gene}}
}
\usage{
TSS.human.NCBI36
}
\description{
TSS annotation for human sapiens (NCBI36) obtained from biomaRt
}
\details{
used in the examples Annotation data obtained by:

mart = useMart(biomart = "ensembl_mart_47", dataset =
"hsapiens_gene_ensembl", archive=TRUE)

getAnnotation(mart, featureType = "TSS")
}
\examples{

data(TSS.human.NCBI36)
slotNames(TSS.human.NCBI36)

}
\keyword{datasets}
