% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.mouse.GRCm38.R
\docType{data}
\name{TSS.mouse.GRCm38}
\alias{TSS.mouse.GRCm38}
\title{TSS annotation data for Mus musculus (GRCm38.p1) obtained from biomaRt}
\format{
GRanges with slot start holding the start position of the gene, slot
end holding the end position of the gene, slot names holding ensembl gene
id, slot seqnames holding the chromosome location where the gene is located
and slot strand holding the strinad information. In addition, the following
variables are included.  \describe{ \item{list("description")}{description
of the gene}}
}
\usage{
TSS.mouse.GRCm38
}
\description{
TSS annotation data for Mus musculus (GRCm38.p1) obtained from biomaRt
}
\details{
Annotation data obtained by:

mart = useMart(biomart = "ensembl", dataset = "mmusculus_gene_ensembl")

getAnnotation(mart, featureType = "TSS")
}
\examples{

data(TSS.mouse.GRCm38)
slotNames(TSS.mouse.GRCm38)

}
\keyword{datasets}
