% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privateUtil.R
\name{TxDb2GR}
\alias{TxDb2GR}
\title{TxDb object to GRanges}
\usage{
TxDb2GR(ranges, feature, OrganismDb)
}
\arguments{
\item{ranges}{an Txdb object}

\item{feature}{feature type, could be geneModel, gene, exon, 
transcript, CDS, fiveUTR, threeUTR, microRNA, and tRNA}

\item{OrganismDb}{org db object}
}
\description{
convert TxDb object to GRanges
}
