% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xget.R
\name{xget}
\alias{xget}
\title{Return the value from a Bimap objects}
\usage{
xget(
  x,
  envir,
  mode,
  ifnotfound = NA,
  inherits,
  output = c("all", "first", "last")
)
}
\arguments{
\item{x, envir, mode, ifnotfound, inherits}{see
\link[AnnotationDbi:Bimap-envirAPI]{mget}}

\item{output}{return the all or first item for each query}
}
\value{
a character vector
}
\description{
Search by name for an Bimap object.
}
\examples{

    library(org.Hs.eg.db)
    xget(as.character(1:10), org.Hs.egSYMBOL)

}
\seealso{
See Also as \code{\link[base:get]{mget}},
\code{\link[AnnotationDbi:Bimap-envirAPI]{mget}}
}
\author{
Jianhong Ou
}
\keyword{misc}
